+++ packages/true_bloc/example/main.dart
@@ -40,11 +40,17 @@ class SimpleBlocObserver extends BlocObserver {
   }
 }
 
-void main() {
+void main() async {
   Bloc.observer = SimpleBlocObserver();
 
   cubitMain();
   blocMain();
+
+  /// Wait for next iteration of the event-loop
+  /// to ensure event has been processed.
+  await Future<void>.delayed(Duration.zero);
+
+  blocCubitMain();
 }
 
 void cubitMain() {
@@ -89,8 +95,53 @@ void blocMain() async {
   await bloc.close();
 }
 
+void blocCubitMain() async {
+  print('----------BLOC+CUBIT----------');
+
+  /// Create a `CounterBloc` instance.
+  final bloc = CounterBlocCubit();
+
+  /// Access the state of the `bloc` via `state`.
+  print(bloc.state); // 0
+
+  /// Interact with the `cubit` to trigger `state` changes.
+  bloc.increment();
+
+  /// Access the new `state`.
+  print(bloc.state); // 1
+
+  /// Interact with the `bloc` to trigger `state` changes.
+  bloc.add(CounterEvent.increment);
+
+  /// Wait for next iteration of the event-loop
+  /// to ensure event has been processed.
+  await Future<void>.delayed(Duration.zero);
+
+  /// Access the new `state`.
+  print(bloc.state); // 2
+
+  /// Interact with the `cubit` to trigger `state` changes.
+  bloc.increment();
+
+  /// Access the new `state`.
+  print(bloc.state); // 3
+
+  /// Interact with the `bloc` to trigger `state` changes.
+  bloc.add(CounterEvent.increment);
+
+  /// Wait for next iteration of the event-loop
+  /// to ensure event has been processed.
+  await Future<void>.delayed(Duration.zero);
+
+  /// Access the new `state`.
+  print(bloc.state); // 4
+
+  /// Close the `bloc` when it is no longer needed.
+  await bloc.close();
+}
+
 /// A `CounterCubit` which manages an `int` as its state.
-class CounterCubit extends Cubit<int> {
+class CounterCubit extends Bloc<Never, int> with EmitterMixin<Never, int> {
   /// The initial state of the `CounterCubit` is 0.
   CounterCubit() : super(0);
 
@@ -120,4 +171,31 @@ class CounterBloc extends Bloc<CounterEvent, int> {
         break;
     }
   }
+}
+
+/// A `CounterBloc` which handles converting `CounterEvent`s into `int`s.
+class CounterBlocCubit extends Bloc<CounterEvent, int>
+    with EmitterMixin<CounterEvent, int> {
+  /// The initial state of the `CounterBloc` is 0.
+  CounterBlocCubit() : super(0);
+
+  @override
+  Stream<int> mapEventToState(CounterEvent event) async* {
+    super.mapEventToState(event);
+
+    switch (event) {
+
+      /// When a `CounterEvent.increment` event is added,
+      /// the current `state` of the bloc is accessed via the `state` property
+      /// and a new state is emitted via `yield`.
+      case CounterEvent.increment:
+        yield state + 1;
+        break;
+    }
+  }
+
+  /// When increment is called, the current state
+  /// of the cubit is accessed via `state` and
+  /// a new `state` is emitted via `emit`.
+  void increment() => emit(state + 1);
 }