+++ packages/true_bloc/test/emitter/counter/flat_map_bloc.dart
@@ -1,0 +1,41 @@
+import 'package:rxdart/rxdart.dart';
+import 'package:true_bloc/bloc.dart';
+
+import '../blocs.dart';
+
+class FlatMapBloc extends Bloc<CounterEvent, int>
+    with EmitterMixin<CounterEvent, int> {
+  FlatMapBloc({this.onTransitionCallback}) : super(0);
+
+  final void Function(Transition<CounterEvent, int>)? onTransitionCallback;
+
+  @override
+  void onTransition(Transition<CounterEvent, int> transition) {
+    super.onTransition(transition);
+    onTransitionCallback?.call(transition);
+  }
+
+  @override
+  Stream<Transition<CounterEvent, int>> transformEvents(
+    Stream<CounterEvent> events,
+    TransitionFunction<CounterEvent, int> transitionFn,
+  ) {
+    return events.flatMap(transitionFn);
+  }
+
+  @override
+  Stream<int> mapEventToState(
+    CounterEvent event,
+  ) async* {
+    super.mapEventToState(event);
+
+    switch (event) {
+      case CounterEvent.decrement:
+        yield state - 1;
+        break;
+      case CounterEvent.increment:
+        yield state + 1;
+        break;
+    }
+  }
+}