+++ packages/true_bloc/test/emitter/simple/simple_bloc.dart
@@ -1,0 +1,19 @@
+import 'dart:async';
+
+import 'package:true_bloc/bloc.dart';
+
+class SimpleBloc extends Bloc<dynamic, String>
+    with EmitterMixin<dynamic, String> {
+  SimpleBloc() : super('');
+
+  @override
+  Stream<String> mapEventToState(dynamic event) async* {
+    super.mapEventToState(event);
+
+    yield 'data';
+  }
+
+  void test() {
+    emit('test');
+  }
+}